Resolve all names that server knows
It is necessary because some devices sometimes
make mDNS requests that includes more than just
the wanted name. Apple does it a lot by including
some annoying bonjour stuff in the middle of a
mDNS request that would be to resolve just a
local name.

This commit fixes that, by supporting this kind
of packet. It will detect some .local(s) that
the server knows and make a response based on
them.

It also gives support to name compression (0xC0).